+++ scripts/eval/benchmark_recon.py
@@ -23,7 +23,7 @@ import pathlib as plib
 from lensless.eval.benchmark import benchmark
 import matplotlib.pyplot as plt
 from lensless import ADMM, FISTA, GradientDescent, NesterovGradientDescent
-from lensless.dataset import DiffuserCamTestDataset
+from lensless.utils.dataset import DiffuserCamTestDataset
 
 try:
     import torch
